# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CartItemAdd": {
        "properties": {
          "color": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required if the product has colours",
            "title": "Color"
          },
          "product_id": {
            "minimum": 1.0,
            "title": "Product Id",
            "type": "integer"
          },
          "quantity": {
            "default": 1,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          },
          "size": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required if the product has sizes",
            "title": "Size"
          }
        },
        "required": [
          "product_id"
        ],
        "title": "CartItemAdd",
        "type": "object"
      },
      "CartItemRead": {
        "properties": {
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "line_total_minor": {
            "title": "Line Total Minor",
            "type": "integer"
          },
          "product": {
            "$ref": "#/components/schemas/ProductSummary"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "size": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "unit_price_minor": {
            "description": "Current product price, in cents",
            "title": "Unit Price Minor",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "product",
          "quantity",
          "size",
          "color",
          "unit_price_minor",
          "line_total_minor"
        ],
        "title": "CartItemRead",
        "type": "object"
      },
      "CartItemUpdate": {
        "properties": {
          "quantity": {
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "quantity"
        ],
        "title": "CartItemUpdate",
        "type": "object"
      },
      "CartRead": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "item_count": {
            "description": "Total units across all lines (the header badge)",
            "title": "Item Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CartItemRead"
            },
            "title": "Items",
            "type": "array"
          },
          "subtotal_minor": {
            "title": "Subtotal Minor",
            "type": "integer"
          },
          "total_minor": {
            "description": "Equals the subtotal: shipping is free, no tax",
            "title": "Total Minor",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "items",
          "item_count",
          "subtotal_minor",
          "total_minor"
        ],
        "title": "CartRead",
        "type": "object"
      },
...
      "ColorRead": {
        "properties": {
          "hex": {
            "title": "Hex",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "hex"
        ],
        "title": "ColorRead",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "LoginIn": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 72,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginIn",
        "type": "object"
      },
...
      "ProductSummary": {
        "properties": {
          "colors": {
            "description": "Colour options; the first is the default",
            "items": {
              "$ref": "#/components/schemas/ColorRead"
            },
            "title": "Colors",
            "type": "array"
          },
          "compare_at_price_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Original price in US cents when the product is on sale",
            "title": "Compare At Price Minor"
          },
          "discount_percent": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Whole-number discount, rounded half up",
            "title": "Discount Percent"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "in_stock": {
            "title": "In Stock",
            "type": "boolean"
          },
          "is_new": {
            "title": "Is New",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price_minor": {
            "description": "Price in US cents",
            "title": "Price Minor",
            "type": "integer"
          },
          "rating_avg": {
            "title": "Rating Avg",
            "type": "number"
          },
          "review_count": {
            "title": "Review Count",
            "type": "integer"
          },
          "sizes": {
            "description": "Size options; the first is the default. Empty if none",
            "items": {
              "type": "string"
            },
            "title": "Sizes",
            "type": "array"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "subtitle": {
            "title": "Subtitle",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "subtitle",
          "price_minor",
          "compare_at_price_minor",
          "discount_percent",
          "is_new",
          "image_url",
          "rating_avg",
          "review_count",
          "in_stock",
          "sizes",
          "colors"
        ],
        "title": "ProductSummary",
        "type": "object"
      },
...
      "RegisterIn": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "first_name": {
            "maxLength": 50,
            "minLength": 1,
            "title": "First Name",
            "type": "string"
          },
          "last_name": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Last Name",
            "type": "string"
          },
          "password": {
            "description": "At least 8 characters with at least one letter and one digit",
            "maxLength": 72,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "first_name",
          "last_name"
        ],
        "title": "RegisterIn",
        "type": "object"
      },
...
      "TokenRead": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserRead"
          }
        },
        "required": [
          "access_token",
          "user"
        ],
        "title": "TokenRead",
        "type": "object"
      },
      "UserRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "first_name",
          "last_name",
          "created_at"
        ],
        "title": "UserRead",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Access token from POST /auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Furniro API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_CREDENTIALS"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error (code VALIDATION_ERROR)"
          }
        },
        "summary": "Log in with email and password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not logged in"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error (code VALIDATION_ERROR)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "The logged-in user",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "EMAIL_ALREADY_REGISTERED"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error (code VALIDATION_ERROR)"
          }
        },
        "summary": "Create an account and log in",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/cart": {
      "delete": {
        "operationId": "clear_cart_api_v1_cart_delete",
        "parameters": [
          {
            "description": "Cart UUID returned by POST /cart",
            "in": "header",
            "name": "X-Cart-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cart UUID returned by POST /cart",
              "title": "X-Cart-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Cart (or item) not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error (code VALIDATION_ERROR)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Empty the cart",
        "tags": [
          "cart"
        ]
      },
      "get": {
        "operationId": "get_cart_api_v1_cart_get",
        "parameters": [
          {
            "description": "Cart UUID returned by POST /cart",
            "in": "header",
            "name": "X-Cart-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cart UUID returned by POST /cart",
              "title": "X-Cart-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Cart (or item) not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error (code VALIDATION_ERROR)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "The cart named by X-Cart-Id",
        "tags": [
          "cart"
        ]
      },
      "post": {
        "operationId": "create_cart_api_v1_cart_post",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error (code VALIDATION_ERROR)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create a guest cart (logged in: returns your cart)",
        "tags": [
          "cart"
        ]
      }
    },
    "/api/v1/cart/items": {
      "post": {
        "operationId": "add_item_api_v1_cart_items_post",
        "parameters": [
          {
            "description": "Cart UUID returned by POST /cart",
            "in": "header",
            "name": "X-Cart-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cart UUID returned by POST /cart",
              "title": "X-Cart-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartItemAdd"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Cart (or item) not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INSUFFICIENT_STOCK"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error (code VALIDATION_ERROR)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add a product (merges with an identical line)",
        "tags": [
          "cart"
        ]
      }
    },
    "/api/v1/cart/items/{item_id}": {
      "delete": {
        "operationId": "remove_item_api_v1_cart_items__item_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Item Id",
              "type": "integer"
            }
          },
          {
            "description": "Cart UUID returned by POST /cart",
            "in": "header",
            "name": "X-Cart-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cart UUID returned by POST /cart",
              "title": "X-Cart-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Cart (or item) not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error (code VALIDATION_ERROR)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove a line",
        "tags": [
          "cart"
        ]
      },
      "patch": {
        "operationId": "update_item_api_v1_cart_items__item_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Item Id",
              "type": "integer"
            }
          },
          {
            "description": "Cart UUID returned by POST /cart",
            "in": "header",
            "name": "X-Cart-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cart UUID returned by POST /cart",
              "title": "X-Cart-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartItemUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Cart (or item) not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INSUFFICIENT_STOCK"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error (code VALIDATION_ERROR)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change a line's quantity",
        "tags": [
          "cart"
        ]
      }
    },
    "/api/v1/cart/merge": {
      "post": {
        "operationId": "merge_cart_api_v1_cart_merge_post",
        "parameters": [
          {
            "description": "Cart UUID returned by POST /cart",
            "in": "header",
            "name": "X-Cart-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cart UUID returned by POST /cart",
              "title": "X-Cart-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not logged in"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error (code VALIDATION_ERROR)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "After login: move the X-Cart-Id guest cart into your cart",
        "tags": [
          "cart"
        ]
      }
    },
...
  }
}